# backend/rest/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API para emissão de Badges",
    "description": "\n## Token de acesso\nEnvia email e senha via Post como o comando de exemplo:\n```bash\ncurl -X POST 'https://{host:port}/user/login' -d \"email=EMAIL&password=SENHA\"\n```\n\nResposta:\n```javascript\n{\n    \"access_token\": \"TOKENDEACESSO\",\n    \"token_type\": \"Bearer\",\n    \"expires_in\": 86400,\n    \"refresh_token\": \"TOKENDEREFRESH\",\n}\n```\n\n## Expiração de Token\nToken de expiração para renovar  por examplo:\n\n```bash\ncurl -X POST 'https://{host:port}/user/refresh' -d \"refreshToken=refresh_token```\n\n"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "any description..."
    }
  },
  "security": [{"AuthorizationHeader": []}],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/badgeclass": {
      "get": {
        "description": "Busca todas as classes",
        "tags":["Badge Classes"],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        },
        "security": [ { "bearerAuth": [] } ]
      },
      "post": {
        "description": "Cria uma nova classe",
        "tags":["Badge Classes"],
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "createdBy": {
                  "example": "string"
                },
                "name": {
                  "example": "string"
                },
                "issuerId": {
                  "example": "string"
                },
                "image": {
                  "example": "string"
                },
                "description": {
                  "example": "string"
                },
                "criteriaUrl": {
                  "example": "string"
                },
                "criteriaNarrative": {
                  "example": "string"
                },
                "alignmentsTargetName": {
                  "example": "string"
                },
                "alignmentsTargetUrl": {
                  "example": "string"
                },
                "alignmentsTargetDescription": {
                  "example": "string"
                },
                "alignmentsTargetFramework": {
                  "example": "string"
                },
                "alignmentsTargetCode": {
                  "example": "string"
                },
                "tags": {
                  "example": "string"
                },
                "expiresAmount": {
                  "example": "string"
                },
                "expiresDuration": {
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        }
      },
      "patch": {
        "description": "Atualiza uma classe",
        "tags":["Badge Classes"],
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "string"
                },
                "name": {
                  "example": "string"
                },
                "description": {
                  "example": "string"
                },
                "criteriaUrl": {
                  "example": "string"
                },
                "criteriaNarrative": {
                  "example": "string"
                },
                "alignmentsTargetName": {
                  "example": "string"
                },
                "alignmentsTargetUrl": {
                  "example": "string"
                },
                "alignmentsTargetDescription": {
                  "example": "string"
                },
                "alignmentsTargetFramework": {
                  "example": "string"
                },
                "alignmentsTargetCode": {
                  "example": "string"
                },
                "tags": {
                  "example": "string"
                },
                "expiresAmount": {
                  "example": "string"
                },
                "expiresDuration": {
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        }
      }
    },
...
    "/assertions": {
      "get": {
        "description": "Busca todas as badges emitidas",
        "tags":["Assertions"],
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          {
          "name": "authorization",
          "in": "header",
          "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        }
      },
      "post": {
        "description": "Emite uma nova badge a um usuário",
        "tags":["Assertions"],
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "issuerId": {
                  "example": "string"
                },
                "badgeclassId": {
                  "example": "string"
                },
                "recipientId": {
                  "example": "string"
                },
                "expires": {
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        }
      }
    },
    "/assertions/{entityId}": {
      "get": {
        "description": "Busca uma badge específica",
        "tags":["Assertions"],
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        }
      },
      "delete": {
        "description": "Excluir uma badge",
        "tags":["Assertions"],
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        }
      }
    },
...
    "/users": {
      "post": {
        "description": "Cria um usuário",
        "security": [ { "bearerAuth": [] } ],
        "tags":["Users"],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "string"
                },
                "email": {
                  "example": "string"
                },
                "type": {
                  "example": "string"
                },
                "image": {
                  "example": "string"
                },
                "password": {
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        }
      },
      "get": {
        "description": "Busca um usuário",
        "tags":["Users"],
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Negado"
          }
        }
      }
    },
...
  },
...
}